Use screen_name instead of user_id
Example:
I will be using @SpaceLaunchNow (user_id= 841696172259606528) as the
example account. The code as is first tries the URI - if a package is
not available to handle the twitter:// URI is available then it falls
back to the web URL.

As is the URI works ONLY with the user_id - however the web url ONLY
works with the screen name.

Example A - URI doesn’t work but weblink does:
twitter://user?user_id= SpaceLaunchNow
https://twitter.com/SpaceLaunchNow

Example B - URI works but weblink doesnt:
twitter://user?user_id=841696172259606528
https://twitter.com/841696172259606528

The obvious solution is to use just the screen_name or just the user_id
for the URI and fallback link. From what I can see the web url only
accepts screen names so changing the URI to screen_name makes more
sense.